+++ Repositories/MicroORM/Memento/CodeCatalog.DDD.Memento/CodeCatalog.DDD.Domain/UseCases/CheckOutOrder.cs
@@ -14,7 +14,7 @@ namespace CodeCatalog.DDD.Domain.UseCases
             _orderRepository = orderRepository;
         }
 
-        public double Checkout(Guid orderId)
+        public decimal Checkout(Guid orderId)
         {
             try
             {
