+++ Repositories/MicroORM/Memento/CodeCatalog.DDD.Memento/CodeCatalog.DDD.Domain/UseCases/ProductRequest.cs
@@ -5,8 +5,8 @@ namespace CodeCatalog.DDD.Domain.UseCases
     public class ProductRequest
     {
         public ProductId ProductId { get; set; }
-        public double Price { get; set; }
-        public double Discount { get; set; }
+        public decimal Price { get; set; }
+        public decimal Discount { get; set; }
         public uint Quantity { get; set; }
     }
 }
